Add ability to copy code snippets to the clipboard
This commit splits the post contents into its own client-side component.
Adds a button on top of every code snippet (visible on hover) that
allows the user to copy the contents of the code into the clipboard. The
code is cleaned up from any annotation markers, as well as any prompt
symbols ($ and #) at the beginnng of each line for console commands.